Separated unit testing from workflow tests.
